Adding queue created attribute, early work to support queue attribute map

+++ sqs/queue.go
@@ -27,6 +27,7 @@ type AWSSqsQueueSpec struct {
 	QueueName string `json:"queueName"`
 	Region    string `json:"region"`
 	QueueUrl  string `json:"queueUrl"`
+  Attributes map[string]*string `json:"attributes"`
 }
 
 type AWSSqsQueueList struct {
@@ -74,9 +75,10 @@ func (c *Controller) AddQueue(obj interface{}) {
 	fmt.Printf("Creating queue %s\n", queue.Spec.QueueName)
 
 	c.SetupService(queue.Spec.Region)
-
-	input := &sqs.CreateQueueInput{
+  fmt.Printf("Result %v\n", queue.Spec.Attributes)
+  input := &sqs.CreateQueueInput{
 		QueueName: &queue.Spec.QueueName,
+    Attributes: queue.Spec.Attributes,
 	}
 
 	response, err := c.svc.CreateQueue(input)
@@ -99,6 +101,10 @@ func (c *Controller) AddQueue(obj interface{}) {
 
 	queueCopy := copiedObject.(*AWSSqsQueue)
 	queueCopy.Spec.QueueUrl = *response.QueueUrl
+  if queueCopy.ObjectMeta.Annotations == nil {
+    queueCopy.ObjectMeta.Annotations = make(map[string]string)
+  }
+  queueCopy.ObjectMeta.Annotations[AWSSqsQueueCRDGroup + "/sqs-autocreated"] = "true"
 
 	err = c.restClient.Put().
 		Name(queue.ObjectMeta.Name).
@@ -124,6 +130,11 @@ func (c *Controller) DeleteQueue(obj interface{}) {
 	input := &sqs.DeleteQueueInput{
 		QueueUrl: &queue.Spec.QueueUrl,
 	}
+  
+  if queue.ObjectMeta.Annotations[AWSSqsQueueCRDGroup + "/sqs-autocreated"] != "true" {
+    fmt.Printf("Refusing to delete %s - queue not created by us", queue.Spec.QueueUrl)
+    return
+  }
 
 	_, err := c.svc.DeleteQueue(input)
 
@@ -136,5 +147,5 @@ func (c *Controller) DeleteQueue(obj interface{}) {
 }
 
 func (c *Controller) UpdateQueue(old, new interface{}) {
-	fmt.Printf("Update not implemented")
+	fmt.Printf("Update not implemented\n")
 }
